# Fix segmenter complexity in resetLevels